% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dic_country}
\alias{dic_country}
\title{Country Dictionary}
\format{A tibble with 502 rows and 3 variables:
\describe{
  \item{con.output}{standard output}
  \item{con.input}{full names abbreviations}
}}
\usage{
data(dic_country)
}
\description{
A list of abbreviations for the United States.
}
\examples{
head(dic_country)

}
\keyword{datasets}
