% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dic_us_dir}
\alias{dic_us_dir}
\title{Directional Dictionary, United States of America}
\format{A tibble with...
\decribe{
  \item{dir.output}{standard directional abbreviations}
  \item{dir.input}{standard full names and directional abbreviations}
}}
\usage{
data(dic_us_dir)
}
\description{
A list of abbreviations for full names for prefix and suffix directionals
   in English.
}
\keyword{datasets}
