% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dic_us_intersect}
\alias{dic_us_intersect}
\title{Intersection Dictionary, United States of America}
\format{A tibble with 5 rows and 2 variables:
\describe{
  \item{intersect.output}{standard intersection identifier}
  \item{intersect.input}{possible intersection identifiers}
}}
\usage{
data(dic_us_intersect)
}
\description{
A list of abbreviations used to identify intersections in the United States.
}
\examples{
head(dic_us_intersect)

}
\keyword{datasets}
