% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dic_us_states}
\alias{dic_us_states}
\title{State Dictionary, United States of America}
\format{A tibble with 124 rows and 2 variables:
\describe{
  \item{state.output}{standard two-letter abbreviations}
  \item{state.input}{standard full names and two-letter abbreviations}
}}
\usage{
data(dic_us_states)
}
\description{
A listing of abbreviations and full names for U.S. states, territories,
    and military "states" used for addressing mail to members of the U.S. armed forces.
    This dictionary includes all territories given two-letter state (or equivalent)
    abbreviations by the U.S. Postal Service.
}
\examples{
head(dic_us_states)

}
\seealso{
\href{https://pe.usps.com/text/pub28/28apb.htm}{U.S. Postal Service, Publication 28, Appendix B}
}
\keyword{datasets}
