% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dic_us_suffix}
\alias{dic_us_suffix}
\title{Street Suffix Dictionary, United States of America}
\format{A tibble with 502 rows and 3 variables:
\describe{
  \item{suf.type}{suffix type}
  \item{suf.output}{standard suffix abbreviation}
  \item{suf.input}{standard full names and suffix abbreviations}
}}
\usage{
data(dic_us_suffix)
}
\description{
A list of abbreviations for full names for street suffix types for the
    United States.
}
\keyword{datasets}
