% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_aptsuffix.R
\name{ifUnit}
\alias{ifUnit}
\title{Parse apartment and unit into standarazed formte}
\usage{
ifUnit(.data, var)
}
\arguments{
\item{.data}{A tbl}

\item{unit}{Name of the variable containing unit/apt}

\item{output}{Optional name for output variable}
}
\description{
\code{pm_aptsuffix.R} identifies aparemnt and unit numbers (i.e. 123 N Washington St apt 3E ) and
parses those directions into their own variable.
}
