% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{pm_all_city}
\alias{pm_all_city}
\title{Does City Name Dictionary Return a Match for All Observations}
\usage{
pm_any_city(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}
}
\value{
If \code{scalar = TRUE}, a single logical scalar is returned that is
    \code{TRUE} if the data contains at least one city name from the given
    dictionary and \code{FALSE} if they do not.
}
\description{
Determine whether the dictionary returns any matches.
}
