% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_all_house}
\alias{pm_all_house}
\title{Do All Addresses Have House Numbers}
\usage{
pm_all_house(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A logical scalar is returned that is \code{TRUE} if all observations contain
    house numbers and \code{FALSE} otherwise.
}
\description{
Determine whether the house number test returns matches for every
    observation.
}
