% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{pm_all_postal}
\alias{pm_all_postal}
\title{Do All Addresses Have Postal Codes}
\usage{
pm_all_postal(.data, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
If \code{scalar = TRUE}, a single logical scalar is returned that is
    \code{TRUE} if all observations contain postal codes and \code{FALSE} otherwise.
}
\description{
Determine whether the postal code test returns matches for every
    observation.
}
