% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{pm_any_dir}
\alias{pm_any_dir}
\title{Do Any Addresses Have Prefix or Suffix Directionals}
\usage{
pm_any_dir(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list of directionals.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains at least
    one directional and \code{FALSE} if they do not.
}
\description{
Determine whether the directional returns any matches.
}
