% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_any_house}
\alias{pm_any_house}
\title{Do Any Addresses Have House Numbers}
\usage{
pm_any_house(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains at least
    one house number and \code{FALSE} if they do not.
}
\description{
Determine whether the house number test returns any matches.
}
