% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{pm_any_state}
\alias{pm_any_state}
\title{Does State Dictionary Return Any Matches}
\usage{
pm_any_state(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
If \code{scalar = TRUE}, a single logical scalar is returned that is
    \code{TRUE} if the data contains at least one state name or abbreviation the given
    dictionary and \code{FALSE} if they do not.
}
\description{
Determine whether the dictionary returns any matches.
}
