% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_checkData.R
\name{pm_checkData}
\alias{pm_checkData}
\title{Check data}
\usage{
pm_checkData(.data)
}
\arguments{
\item{.data}{A tbl}
}
\description{
Since \code{postmastr} uses a set of default variable names for returning parsed output,
this function can be used to ensure that there will be no conflicts between user data
and \code{posmastr}.
}
