% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{pm_city_detect}
\alias{pm_city_detect}
\title{Detect Presence of City Name in Address}
\usage{
pm_city_detect(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}
}
\value{
A tibble with a new logical variable \code{pm.hasCity} that is
    \code{TRUE} if a city name from the given dictionary is found in the
    at the end the address and \code{FALSE} otherwise.
}
\description{
Determine the presence of city names in a string.
}
