% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{pm_city_none}
\alias{pm_city_none}
\title{Return Only Unmatched Observations From pm_has_city}
\usage{
pm_city_none(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}
}
\value{
A tibble containing only observations that were not found in
    the dictionary. The variable created by \link{pm_has_city},
    \code{pm.hasCity}, is removed.
}
\description{
Automatically subset the results of \link{pm_has_city} to
   return only observations that were not found in the dictionary.
}
