% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{pm_country_all}
\alias{pm_country_all}
\title{Does Country Dictionary Return a Match for All Observations}
\usage{
pm_country_all(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for countries. If none is specified, the full default
country dictionary will be used.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains a country
    name or abbreviation for every observation in the data set and \code{FALSE} otherwise.
}
\description{
Determine whether the country dictionary returns matches for all observations.
}
