% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{pm_country_any}
\alias{pm_country_any}
\title{Does Country Dictionary Return Any Matches}
\usage{
pm_country_any(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for countries. If none is specified, the full default
country dictionary will be used.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains at
    least one country name or abbrevation in the given dictionary and \code{FALSE}
    if they do not.
}
\description{
Determine whether the country dictionary returns any matches.
}
