% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{pm_country_detect}
\alias{pm_country_detect}
\title{Detect Presence of Country}
\usage{
pm_country_detect(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for countries. If none is specified, the full default
country dictionary will be used.}
}
\value{
A tibble with a new logical variable \code{pm.hasCountry} that is
    \code{TRUE} if a country name or abbreviation from the given dictionary is
    found at the end of the address and \code{FALSE} otherwise.
}
\description{
Determine the presence of country names or abbreviations
    at the end of a string.
}
