% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{pm_country_none}
\alias{pm_country_none}
\title{Return Only Unmatched Observations From pm_country_detect}
\usage{
pm_country_none(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for countries. If none is specified, the full default
country dictionary will be used.}
}
\description{
Automatically subset the results of \link{pm_country_detect} to
   return only observations that were not found in the dictionary.
}
