% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{pm_country_std}
\alias{pm_country_std}
\title{Standardize Parsed Countries}
\usage{
pm_country_std(.data, var, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable that may contain countries}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for countries. If none is specified, the full default
country dictionary will be used.}
}
\description{
Convert countries to USPS preferred two-letter abbreviation.
}
