% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{pm_country_trim}
\alias{pm_country_trim}
\title{Trim Country}
\usage{
pm_country_trim(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for countries. If none is specified, the full default
country dictionary will be used.}
}
\description{
Remove a country from an address without parsing. These data
    should be at the end of the string (i.e. the last word or words).
}
