% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{pm_dictionary}
\alias{pm_dictionary}
\title{Create Dictionary Objects}
\usage{
pm_dictionary(locale = "us", type, append, filter)
}
\arguments{
\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}

\item{type}{A string indicating the grammatical address element the dictionary
should represent. Current options are \code{"state"} and \code{"city"}.}

\item{append}{An optional dictionary appendix object created with \code{\link{pm_append}}}

\item{filter}{An optional character scalar or vector with output elements that should
be retained.}
}
\value{
A \code{postmastr} dictionary object, which will always include an input column
    of possible terms for the given grammatical address element and an output column
    with the desired output for each input.
}
\description{
This function allows for the creation of dictionary objects that are
    either optional or required elements of other \code{postmastr} functions.
}
