% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{pm_evaluate}
\alias{pm_evaluate}
\title{Evaluate Address Type}
\usage{
pm_evaluate(.data)
}
\arguments{
\item{.data}{A source tibble that has already had identification
numbers added using \link{pm_identify}.}
}
\value{
A tibble with the frequency and percent of values in the \code{pm.type} variable.
}
\description{
Prints a tibble with the results of \code{\link{pm_identify}}'s assessment
    of address types.
}
