% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{pm_has_city}
\alias{pm_has_city}
\title{Detect Presence of City Name}
\usage{
pm_has_city(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object (\code{pm_subset})}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}
}
\value{
A tibble with a new logical variable \code{pm.isCity} that is
    \code{TRUE} if a city name is found in the address and \code{FALSE}
    otherwise.
}
\description{
Determine the presence of city names in a string.
}
