% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_has_house}
\alias{pm_has_house}
\title{Detect Presence of House Numbers}
\usage{
pm_has_house(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble with a new logical variable \code{pm.hasHouse} that is
    \code{TRUE} if a house number is found in the first word of the address
    and \code{FALSE} otherwise.
}
\description{
Determine the presence of house numbersin a string.
}
