% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{pm_has_state}
\alias{pm_has_state}
\title{Detect Presence of State Name or Abbreviation}
\usage{
pm_has_state(.data, dictionary, scalar = TRUE, locale = "us")
}
\arguments{
\item{.data}{A tbl or data frame}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for states. If none is provided, the \code{states}
object will be used as the default directory.}

\item{scalar}{If \code{TRUE}, a single logical scalar is returned; otherwise if
\code{FALSE}, a logical vector is returned.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new logical variable \code{pm.isState} that is
    \code{TRUE} if a state name or abbreviation is found in the address
    and \code{FALSE} otherwise.

If \code{scalar = TRUE}, a single logical scalar is returned that is
    \code{TRUE} if the data contain statenames or abbreviations and \code{FALSE}
    if they do not. If \code{scalar = FALSE} a tibble with a new logical variable
    \code{pm.isState} that is \code{TRUE} if a state name or abbreviation is found
    in the last word of the address and \code{FALSE} otherwise.
}
\description{
Determine the presence of state names or abbreviations
    at the end of a string.
}
