% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{pm_has_street_dir}
\alias{pm_has_street_dir}
\title{Detect Presence of Prefix or Suffix Directionals}
\usage{
pm_has_street_dir(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list of directionals.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A tibble with a new logical variable \code{pm.hasDir} that is
    \code{TRUE} if a directional is found in the first or last word of
    the address and \code{FALSE} otherwise.
}
\description{
Determine the presence of house numbersin a string.
}
