% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_house_none}
\alias{pm_house_none}
\title{Return Only Unmatched Observations From pm_house_detect}
\usage{
pm_house_none(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble containing only observations that were not found matched
    using the house number test. The variable created by \link{pm_house_detect},
    \code{pm.hasHouse}, is removed.
}
\description{
Automatically subset the results of \link{pm_house_detect} to
   return only observations that were not found to include a house number
}
