% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_houseAlpha.R
\name{pm_houseAlpha}
\alias{pm_houseAlpha}
\title{Parse Alphanumeric Values in House Number}
\source{
See \href{https://pe.usps.com/text/pub28/28ape_003.htm}{USPS Publication 28 -
    Postal Addressing Standards, Section E12}
}
\usage{
pm_houseAlpha(.data, houseNum, alphaVar)
}
\arguments{
\item{.data}{A tbl}

\item{houseNum}{A vector containing house numbers to be parsed}

\item{alphaVar}{Optional name for output variable}
}
\value{
A tibble with the alphanumeric value, which is named \code{houseAlpha} by default
    but can be altered with the \code{alphaVar} argument.
}
\description{
\code{pm_houseAlpha} parses the house number for included alphanumeric ranges. This function
is in development - it currently only identifies "F" (front) and "R" (rear) addresses.
}
