% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseAlpha.R
\name{pm_houseAlpha_all}
\alias{pm_houseAlpha_all}
\title{Do All Addresses Have Alphanumeric House Numbers}
\usage{
pm_houseAlpha_all(.data, position = "any", locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{position}{A string indicating where letters are located; one of either
\code{"any"} (anywhere in the house number), \code{"front"} (initial letters
of the house number), \code{"end"} (last letters of the house number),
or \code{"middle"} (in the middle but not at the ends).}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if all observations contain
    alphanumeric house number and \code{FALSE} otherwise.
}
\description{
Determine whether the alphanumeric house number test returns matches for every
    observation.
}
