% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseAlpha.R
\name{pm_houseAlpha_detect}
\alias{pm_houseAlpha_detect}
\title{Detect Presence of Alphanumeric House Numbers}
\usage{
pm_houseAlpha_detect(.data, position = "any", locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{position}{A string indicating where letters are located; one of either
\code{"any"} (anywhere in the house number), \code{"front"} (initial letters
of the house number), \code{"end"} (last letters of the house number),
or \code{"middle"} (in the middle but not at the ends).}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new logical variable beginning with \code{pm.hasAlpha.}
    that is \code{TRUE} if letters are detected in the specified position and
    \code{FALSE} otherwise. If \code{pm.hasAlpha.} has a \code{a} suffix, it
    corresponds to the \code{"any"} position. If \code{pm.hasAlpha.} has a \code{f}
    suffix, it corresponds to the \code{"front"} position. If \code{pm.hasAlpha.}
    has a \code{e} suffix, it corresponds to the \code{"end"} position. Finally,
    If \code{pm.hasAlpha.} has a \code{m} suffix, it corresponds to the \code{"middle"}
    position. If a \code{NA} value is returned, it is because the house number is an
    address range.
}
\description{
Detect the presence of U.S. alphanumeric house numbers
   in a string address (i.e. 123A Main St).
}
