% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseFrac.R
\name{pm_houseFrac_detect}
\alias{pm_houseFrac_detect}
\title{Detect Presence of Fractional House Numbers}
\usage{
pm_houseFrac_detect(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble with a new logical variable \code{pm.hasHouse} that is
    \code{TRUE} if a house number is found in the first word of the address
    and \code{FALSE} otherwise.
}
\description{
Determine the presence of fractional house numbers in a string.
}
\details{
A fractional house number is used in some parts of the United States.
   Fractional house numbers typically look like \code{123 1/2 Main St}.
   The U.S.P.S allows any fraction, though \code{1/2} appears commonly.
}
