% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_houseFrac_none}
\alias{pm_houseFrac_none}
\title{Return Only Unmatched Observations From pm_houseFrac_detect}
\usage{
pm_houseFrac_none(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble containing only observations that were not found matched
    using the fractional house number test. The variable created by
    \link{pm_houseFrac_detect}, \code{pm.hasHouseFrac}, is removed.
}
\description{
Automatically subset the results of \link{pm_houseFrac_detect} to
   return only observations that were not found to include a fractional house number.
}
