% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseFrac.R
\name{pm_houseFrac_parse}
\alias{pm_houseFrac_parse}
\title{Parse Fractional House Numbers}
\usage{
pm_houseFrac_parse(.data, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new column \code{pm.houseFrac} that contains the fractional house number.
    If a house number is not detected in the string, a value of \code{NA} will be
    returned.
}
\description{
Create a new column containing fractional house number data.
}
