% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_houseNum.R
\name{pm_houseNum}
\alias{pm_houseNum}
\title{Parse House Number}
\usage{
pm_houseNum(.data, address, numVar)
}
\arguments{
\item{.data}{A tbl}

\item{address}{Name of the street address variable containing house numbers}

\item{numVar}{Optional root name for output variable(s)}
}
\value{
A tibble with the house number and, if needed in the case of hyphenated address ranges,
   variables representing the upper and lower bounds of the address range. These variables all
   will share the same root, which is \code{houseNum} by default but can be altered with the
   \code{numVar} argument.
}
\description{
Parse House Number
}
\details{
\code{pm_houseNum} parses the house number from a given address string structured in
the following way - "123 Main St". This function can also process address ranges that
are structured in the following manner - "123-125 Main St"
}
