% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseRange.R
\name{pm_houseRange_all}
\alias{pm_houseRange_all}
\title{Do All Addresses Have House Ranges}
\usage{
pm_houseRange_all(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A logical scalar is returned that is \code{TRUE} if all observations contain
    a house range and \code{FALSE} otherwise.
}
\description{
Determine whether all addresses have a house range.
}
\details{
A house range is used in some parts of the United States.
   House ranges typically look like \code{121-123 Main St}.
}
