% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseRange.R
\name{pm_houseRange_any}
\alias{pm_houseRange_any}
\title{Do Any Addresses Have House Ranges}
\usage{
pm_houseRange_any(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains at least
    house range and \code{FALSE} if they do not.
}
\description{
Determine whether the house range test returns any matches.
}
\details{
A house range is used in some parts of the United States.
   House ranges typically look like \code{121-123 Main St}.
}
