% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseRange.R
\name{pm_houseRange_match}
\alias{pm_houseRange_match}
\title{Return Only Matched Observations From pm_houseRange_detect}
\usage{
pm_houseRange_match(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble containing only observations that were matched
    using the house range test. The variable created by
    \link{pm_houseRange_detect}, \code{pm.hasHouseRange}, is removed.
}
\description{
Automatically subset the results of \link{pm_houseRange_detect} to
   return only observations that were found to include a house range. This functionality
   is only available for house range matching right now because ranges can contain
   complex alphanumeric values that need manual cleaning.
}
\details{
A house range is used in some parts of the United States.
   House ranges typically look like \code{121-123 Main St}.
}
