% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseRange.R
\name{pm_houseRange_none}
\alias{pm_houseRange_none}
\title{Return Only Unmatched Observations From pm_houseRange_detect}
\usage{
pm_houseRange_none(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble containing only observations that were not matched
    using the house range test. The variable created by
    \link{pm_houseRange_detect}, \code{pm.hasHouseRange}, is removed.
}
\description{
Automatically subset the results of \link{pm_houseRange_detect} to
   return only observations that were not found to include a house range.
}
\details{
A house range is used in some parts of the United States.
   House ranges typically look like \code{121-123 Main St}.
}
