% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseSuf.R
\name{pm_houseSuf_all}
\alias{pm_houseSuf_all}
\title{Does House Suffix Dictionary Return a Match for All Observations}
\usage{
pm_houseSuf_all(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_append} to be used
as a master list for house suffixes.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains a house suffix
    for every observation in the data set and \code{FALSE} otherwise.
}
\description{
Determine whether the street suffix dictionary returns matches for all observations.
}
