% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseSuf.R
\name{pm_houseSuf_any}
\alias{pm_houseSuf_any}
\title{Does House Suffix Dictionary Return Any Matches}
\usage{
pm_houseSuf_any(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_append} to be used
as a master list for house suffixes}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains at
    least one house suffix in the given dictionary and \code{FALSE} if they do not.
}
\description{
Determine whether the street suffix dictionary returns any matches.
}
