% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseSuf.R
\name{pm_houseSuf_detect}
\alias{pm_houseSuf_detect}
\title{Detect Presence of House Suffix}
\usage{
pm_houseSuf_detect(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_append} to be used
as a master list for house suffixes}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A tibble with a new logical variable \code{pm.hasHouseSuf} that is
    \code{TRUE} if a house suffix from the given dictionary is found at the beginning
    of the address and \code{FALSE} otherwise.
}
\description{
Determine the presence of street suffix names or abbreviations
    at the end of a string.
}
