% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{pm_houseSuf_none}
\alias{pm_houseSuf_none}
\title{Return Only Unmatched Observations From pm_houseSuf_detect}
\usage{
pm_houseSuf_none(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_append} to be used
as a master list for house suffixes}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\description{
Automatically subset the results of \link{pm_houseSuf_detect} to
   return only observations that were not found in the dictionary.
}
