% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseSuf.R
\name{pm_houseSuf_parse}
\alias{pm_houseSuf_parse}
\title{Parse House Suffix}
\usage{
pm_houseSuf_parse(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_append} to be used
as a master list for house suffixes}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\description{
Parse a house suffix from a string. These data
    should be at the beginning of the string (i.e. the first word).
}
