% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{pm_houseSuf_std}
\alias{pm_houseSuf_std}
\title{Standardize Parsed House Suffix Names}
\usage{
pm_houseSuf_std(.data, var, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable that may contain street suffix values}

\item{dictionary}{A tbl created with \code{pm_append} to be used
as a master list for house suffixes}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A tibble with an updated variable that contains standardized house suffix values.
}
\description{
Convert house suffix values to desired outputs
}
