% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{pm_intersect_all}
\alias{pm_intersect_all}
\title{Does Intersection Dictionary Return a Match for All Observations}
\usage{
pm_intersect_all(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable containing address data to be parsed}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list intersection values.}
}
\value{
A single logical scalar is returned that is
    \code{TRUE} if the data contains at least one intersection from the given
    dictionary and \code{FALSE} if they do not.
}
\description{
Determine whether the intersection dictionary returns any matches.
}
