% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{pm_intersect_detect}
\alias{pm_intersect_detect}
\title{Detect Presence of Intersection in Address}
\usage{
pm_intersect_detect(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable containing address data to be parsed}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list intersection values.}
}
\value{
A tibble with a new logical variable \code{pm.hasIntersect} that is
    \code{TRUE} if an intersection from the given dictionary is found in the
    address and \code{FALSE} otherwise.
}
\description{
Determine the presence of intersection in a string.
}
