% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{pm_intersect_longer}
\alias{pm_intersect_longer}
\title{Convert Intersections to Long Form}
\usage{
pm_intersect_longer(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for intersection operators.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\description{
Split intersections into x and y streets.
}
