% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{pm_intersect_none}
\alias{pm_intersect_none}
\title{Return Only Unmatched Observations From pm_intersect_detect}
\usage{
pm_intersect_none(.data, dictionary)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for intersections.}

\item{var}{A character variable containing address data to be parsed}
}
\value{
A tibble containing only observations that were not found in
    the dictionary. The variable created by \link{pm_intersect_detect},
    \code{pm.hasIntersect}, is removed.
}
\description{
Automatically subset the results of \link{pm_intersect_detect} to
   return only observations that were not found in the dictionary.
}
