% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{pm_is_subset}
\alias{pm_is_subset}
\title{Validate postmastr Object}
\usage{
pm_is_subset(obj)
}
\arguments{
\item{obj}{Object to test}
}
\value{
A logical scalar that is \code{TRUE} if the given object is of class
    \code{pm_subset}; it will return \code{FALSE} otherwise.
}
\description{
This function tests to see whether an object is of class \code{pm_subset}.
    It is used as part of the parsing functions, and is exported so that it can be used
    interactively as well.
}
