% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{pm_isCity}
\alias{pm_isCity}
\title{Detect Presence of City Name}
\usage{
pm_isCity(.data, var, dictionary)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable that may contain city names}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}
}
\value{
A tibble with a new logical variable \code{pm.isCity} that is
    \code{TRUE} if a city name is found in the address and \code{FALSE}
    otherwise.
}
\description{
Determine the presence of city names in a string.
}
