% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{pm_isZip}
\alias{pm_isZip}
\title{Detect Presence of Zip Codes}
\usage{
pm_isZip(.data, var)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable that may contain zip codes}
}
\value{
A tibble with a new logical variable \code{pm.isZip} that is
    \code{TRUE} if a zip-code is found in the last word of the address
    and \code{FALSE} otherwise.
}
\description{
Determine the presence of U.S. zip-codes in a string. This will
    identify both five digit zip-codes (e.g. \code{63108}) as well as zip+4
    codes (e.g. \code{63108-3412}). The zip-code must be the final word in
    the string to return \code{TRUE}.
}
