% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_no_house}
\alias{pm_no_house}
\title{Return Only Unmatched Observations From pm_has_house}
\usage{
pm_no_house(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble containing only observations that were not found matched
    using the house number test. The variable created by \link{pm_has_house},
    \code{pm.hasHouse}, is removed.
}
\description{
Automatically subset the results of \link{pm_has_house} to
   return only observations that were not found to include a house number
}
