% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_no_houseFrac}
\alias{pm_no_houseFrac}
\title{Return Only Unmatched Observations From pm_has_houseFrac}
\usage{
pm_no_houseFrac(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble containing only observations that were not found matched
    using the fractional house number test. The variable created by
    \link{pm_has_houseFrac}, \code{pm.hasHouseFrac}, is removed.
}
\description{
Automatically subset the results of \link{pm_has_houseFrac} to
   return only observations that were not found to include a fractional house number.
}
