% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{pm_no_postal}
\alias{pm_no_postal}
\title{Return Only Unmatched Observations From pm_has_postal}
\usage{
pm_no_postal(.data,  locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A tibble containing only observations that were not found in
    the dictionary. The variable created by \link{pm_has_postal},
    \code{pm.hasZip}, is removed.
}
\description{
Automatically subset the results of \link{pm_has_postal} to
   return only observations that were not found to include a postal (zip) code.
}
