% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pm_parse}
\alias{pm_parse}
\title{Parse Street Address}
\usage{
pm_parse(.data, style, locale = "us", newVar, dirDictionary,
  suffixDictionary, cityDictionary, stateDictionary)
}
\description{
Parse Street Address
}
