% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{pm_parse_houseFrac}
\alias{pm_parse_houseFrac}
\title{Parse Fractional House Numbers}
\usage{
pm_parse_houseFrac(.data)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}
}
\value{
A tibble with a new column \code{pm.houseFrac} that contains the fractional house number.
    If a house number is not detected in the string, a value of \code{NA} will be
    returned.
}
\description{
Create a new column containing fractional house number data.
}
