% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{pm_parse_postal}
\alias{pm_parse_postal}
\title{Parse Postal Codes}
\usage{
pm_parse_zip(.data, locale = "us")
}
\arguments{
\item{.data}{A postmastr object (\code{pm_subset})}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new column \code{pm.zip} that contains the zip-code.
    If a postal code is not detected in the string, a value of \code{NA} will be
    returned. If the "zip+4" formatting is detected in the string, a second column
    named \code{pm.zip4} will be returned with the carrier route parsed out of the
    five-digit postal code.
}
\description{
Create a new column containing postal code data.
}
