% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{pm_parse_street_dir}
\alias{pm_parse_street_dir}
\title{Parse Prefix and Suffix Directionals}
\usage{
pm_parse_street_dir(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for states. If none is provided, the \code{states}
object will be used as the default directory.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new character variable \code{pm.preDir} that contains
    the abbreviation for the given directional for any prefix directional
    and a second new character variable \code{pm.sufDir} that contains the
    abbreviation for the given directional for any suffix directional.
    The use of abbrevations follows USPS addressing standards. If a prefix
    or suffix direction is not detected in the string, a value of \code{NA}
    will be returned. If no prefix directions are found in the data at all,
    that column will not be returned; the same is true for suffix directions.
}
\description{
Parse a prefix or suffix directional from a string. These data
    should be at the beginning or end of the string (i.e. the first/last word or two).
}
