% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{pm_parse_zip}
\alias{pm_parse_zip}
\title{Parse Zip-Codes}
\usage{
pm_parse_zip(.data, var, locale = "us")
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable that may contain zip codes}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new column \code{pm.zip} that contains the zip-code.
    If a zip-code is not detected in the string, a value of \code{NA} will be
    returned. If it does not yet exist, a copy of the address variable will
    be created in \code{pm.address} and returned with zip-codes removed.
}
\description{
Create a new column containing zip-code data.
}
