% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{pm_parseCity}
\alias{pm_parseCity}
\title{Parse City Names}
\usage{
pm_parseCity(.data, var, dictionary)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable that may contain city names}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}
}
\value{
A tibble with a new character variable \code{pm.city} that contains
    the city name. If a city name is not detected in the string, a value
    of \code{NA} will be returned. If it does not yet exist, a copy of the
    address variable will be created in \code{pm.address} and returned with
    state name or abbreviation removed.
}
\description{
Parse a city name or abbreviation from a string. These data
    should be at the end of the string (i.e. the last several words). If a
    state name or abbrevation follows the city, use \link{pm_parseState} first
    to remove those data from \code{pm.address}. Likewise, if a
    zip-code follows a name, use \link{pm_parseZip} first to remove those
    data from \code{pm.address}.
}
