% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{pm_parseZip}
\alias{pm_parseZip}
\title{Parse Zip-Codes}
\usage{
pm_parseZip(.data, var)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable that may contain zip codes}
}
\value{
A tibble with a new column \code{pm.zip} that contains the zip-code.
    If a zip-code is not detected in the string, a value of \code{NA} will be
    returned. If it does not yet exist, a copy of the address variable will
    be created in \code{pm.address} and returned with zip-codes removed.
}
\description{
Create a new column containing zip-code data.
}
