% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{pm_postal_any}
\alias{pm_postal_any}
\title{Do Any Addresses Have Postal Codes}
\usage{
pm_postal_any(.data, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains at least
    one postal code and \code{FALSE} if they do not.
}
\description{
Determine whether the postal code test returns any matches.
}
