% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{pm_postal_detect}
\alias{pm_postal_detect}
\title{Detect Presence of Postal Codes}
\usage{
pm_postal_detect(.data, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new logical variable \code{pm.hasZip} that is
    \code{TRUE} if a zip-code is found in the last word of the address
    and \code{FALSE} otherwise.
}
\description{
Determine the presence of U.S. zip-codes in a string. This will
    identify both five digit zip-codes (e.g. \code{63108}) as well as zip+4
    codes (e.g. \code{63108-3412}). The zip-code must be the final word in
    the string to return \code{TRUE}.
}
