% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pm_rebuild}
\alias{pm_rebuild}
\title{Re-construct Street Addressed from Parsed Elements}
\usage{
pm_rebuild(.data, output, new_address, include_commas = FALSE, include_units = TRUE,
   keep_parsed = "no", side = "right", left_vars, keep_ids = FALSE, locale = "us")
}
\arguments{
\item{.data}{An object with raw and parsed data created by \code{\link{pm_rebuild}}}

\item{output}{Describes the format of the output address. One of either \code{"full"} or \code{"short"}.
A short address contains, at the most, a house number, street directionals, a street name,
a street suffix, and a unit type and number. A full address contains all of the selements of a
short address as well as, at the most, a city, state, and postal code.}

\item{new_address}{Optional; name of new variable to store rebuilt address in. If not specified,
the re-build addressed will be stored in \code{pm.address}.}

\item{include_commas}{A logical scalar; if \code{TRUE}, a comma is added both before and after the city
name in rebuild addresses. If \code{FALSE} (default), no punctuation is added.}

\item{include_units}{A logical scalar; if \code{TRUE} (default), unit data will be included in the
re-constructed address. If \code{FALSE}, units will not be included.}

\item{keep_parsed}{Character string; if \code{"yes"}, all parsed elements will be
added to the source data after replacement. If \code{"limited"}, only the \code{pm.city},
\code{pm.state}, and postal code variables will be retained. Otherwise, if \code{"no"},
only the rebuilt address will be added to the source data (default).}

\item{side}{One of either \code{"left"} or \code{"right"} - should parsed data be placed to the left
or right of the original data? Placing data to the left may be useful in particularly wide
data sets.}

\item{left_vars}{A character scalar or vector of variables to place on the left-hand side of
the output when \code{side} is equal to \code{"middle"}.}

\item{keep_ids}{Logical scalar; if \code{TRUE}, the identification numbers
will be kept in the source data after replacement. Otherwise, if \code{FALSE},
they will be removed (default).}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\description{
Create a single address from parsed components.
}
\details{
Re-constructed street addresses will can be either \code{"full"} or \code{"short"} depending
    on the \code{output} parameter's argument. Intersections will always be re-constructed in full.
}
