% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{pm_state_all}
\alias{pm_state_all}
\title{Does State Dictionary Return a Match for All Observations}
\usage{
pm_state_all(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains a state
    name or abbreviation for every observation in the data set and \code{FALSE} otherwise.
}
\description{
Determine whether the dictionary returns any matches.
}
