% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{pm_state_detect}
\alias{pm_state_detect}
\title{Detect Presence of State Name or Abbreviation}
\usage{
pm_state_detect(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for states. If none is provided, the \code{states}
object will be used as the default directory.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A tibble with a new logical variable \code{pm.hasState} that is
    \code{TRUE} if a state name or abbreviation from the given dictionary is found
    at the end of the address and \code{FALSE} otherwise.
}
\description{
Determine the presence of state names or abbreviations
    at the end of a string.
}
