% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{pm_std_states}
\alias{pm_std_states}
\title{Standardize Parsed State Names}
\usage{
pm_std_states(.data, var, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable that may contain city names}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for states. If none is provided, the \code{states}
object will be used as the default directory.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with an updated variable that contains the two-letter abbreviation
    for the given U.S. state. This follows USPS addressing standards, which require
    that the state abbreviation be used.
}
\description{
Convert state names to the USPS approved two-letter abbreviation.
}
