% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{pm_stdState}
\alias{pm_stdState}
\title{Standardize Parsed State Names}
\usage{
pm_stdState(.data, var, dictionary)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable that may contain city names}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for states. If none is provided, the \code{states}
object will be used as the default directory.}
}
\value{
A tibble with an updated variable that contains the two-letter abbreviation
    for the given U.S. state. This follows USPS addressing standards, which require
    that the state abbreviation be used.
}
\description{
Convert state names to the USPS approved two-letter abbreviation.
}
