% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_stdStreet.R
\name{pm_stdStreet}
\alias{pm_stdStreet}
\title{Standardized Street Names}
\usage{
pm_stdStreet(.data, stName, output)
}
\arguments{
\item{.data}{A tbl}

\item{stName}{Name of the street name variable}

\item{output}{Optional name of output variable}
}
\description{
\code{pm_stdStreet} standardizes the spelling and formatting of street names. Names like "Second" are
consistently converted to "2nd".
}
