% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_stdSuffix.R
\name{pm_stdSuffix}
\alias{pm_stdSuffix}
\title{Standardized Street Suffixes}
\usage{
pw_stSuffix(.data, suffix, overwrite = TRUE, newSuffix)
}
\arguments{
\item{.data}{A tbl}

\item{suffix}{A character vector within \code{.data} that contains street suffixes}

\item{overwrite}{A logical scalar. Should the output overwrite the given variable?}

\item{newSuffix}{A name for a new vector to be created if \code{overwrite = FALSE}}
}
\value{
\code{pw_stSuffix} returns a tibble with the requested output - either the suffix variable has
    been overwritten or a new variable with corrected suffix abbreviations has been added.
}
\description{
\code{pw_stSuffix} standardizes a given set of street suffix values to the USPS preferred suffix abbreviation.
}
\note{
\code{pw_stSuffix} requires that a number variable names be unused in the original data - \code{suf_com},
    \code{suf_cor}, \code{suf_pri} and \code{suf_std}. If these names are present, \code{pw_stSuffix} will
    return an error. These variables are created temporarily as part of the matching process.
}
