% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_stPrefixDir.R
\name{pm_stPrefixDir}
\alias{pm_stPrefixDir}
\title{Parse Prefix Direction}
\usage{
pm_stPrefixDir(.data, stName, std = TRUE, output)
}
\arguments{
\item{.data}{A tbl}

\item{std}{A logical scalar. Should directional values be standardized?}

\item{output}{Optional name for output variable}

\item{address}{Name of the street address variable containing prefix directions}
}
\description{
\code{pm_stSuffixDir} identifies streets with suffix directions (i.e. 123 N Washington St) and
parses those directions into their own variable.
}
