% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street.R
\name{pm_street_parse}
\alias{pm_street_parse}
\title{Parse Street Names}
\usage{
pm_street_parse(.data, dictionary, ordinal = TRUE, drop = TRUE, locale = "us")
}
\arguments{
\item{.data}{A \code{postmastr} object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_append} to be used to standardize
specific street names.}

\item{ordinal}{A logical scalar; if \code{TRUE}, street names that contain numeric words values
(i.e. "Second") will be converted and standardized to ordinal values (i.e. "2nd"). The
default is \code{TRUE} because it returns much more compact clean addresses (i.e.
"168th St" as opposed to "One Hundred Sixty Eigth St").}

\item{drop}{A logical scalar; if \code{TRUE}, the \code{pm.address} variable will
be dropped from the \code{postmastr} object.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with a new character variable \code{pm.street} that contains
    the two-letter abbreviation for the given U.S. state. Variables are automatically
    re-ordered, so the new vector will not be in the last position of the tibble.
}
\description{
Converts the remaining text of \code{pm.address} to title case and stores
    it in a new variable named \code{pm.street}.
}
\details{
This is typically the last function to be executed before rebuilding and replacing.
}
