% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street.R
\name{pm_street_std}
\alias{pm_street_std}
\title{Standardize Street Names}
\usage{
pm_street_std(.data, var, dictionary, ordinal = TRUE, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable that may contain street suffixes}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for street suffixes.}

\item{ordinal}{A logical scalar; if \code{TRUE}, street names that contain numeric words values
(i.e. "Second") will be converted and standardized to ordinal values (i.e. "2nd"). The
default is \code{TRUE} because it returns much more compact clean addresses (i.e.
"168th St" as opposed to "One Hundred Sixty Eigth St").}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\description{
Standardize street names by converting to title case, removing punctuation,
    and optionally applying ordinal conversion as well as a dictionary to the data.
}
