% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{pm_streetDir_all}
\alias{pm_streetDir_all}
\title{Do All Addresses Have Prefix or Suffix Directionals}
\usage{
pm_streetDir_all(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; A tbl created with \code{pm_dictionary} to be used
as a master list of directionals. If none is provided, the \code{dic_us_dir}
object will be used as the default dictionary when \code{locale = "us"}.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if all observations contain
    directionals and \code{FALSE} otherwise.
}
\description{
Determine whether the directional test returns matches for every
    observation.
}
