% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{pm_streetDir_detect}
\alias{pm_streetDir_detect}
\title{Detect Presence of Prefix or Suffix Directionals}
\usage{
pm_streetDir_detect(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list of directionals.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A tibble with a new logical variable \code{pm.hasDir} that is
    \code{TRUE} if a directional is found in the first or last word of
    the address and \code{FALSE} otherwise.
}
\description{
Determine the presence of house numbersin a string.
}
\details{
If a street name is also a directional, like \code{North Ave}, it will be
    identified as such. See \link{pm_streetDir_parse} and \link{pm_streetSuf_parse}
    for additional details on how these addresses are handled.
}
