% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{pm_streetDir_none}
\alias{pm_streetDir_none}
\title{Return Only Unmatched Observations From pm_has_dir}
\usage{
pm_streetDir_none(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for cities.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A tibble containing only observations that were not found in
    the dictionary. The variable created by \link{pm_streetDir_detect},
    \code{pm.hasDir}, is removed.
}
\description{
Automatically subset the results of \link{pm_streetDir_detect} to
   return only observations that were not found in the dictionary.
}
