% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{pm_streetDir_std}
\alias{pm_streetDir_std}
\title{Standardize Parsed State Names}
\usage{
pm_streetDir_std(.data, var, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable that may contain directionals}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for directionals. If none is provided, the \code{dic_us_dir}
object will be used as the default directory.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A tibble with an updated variable that contains the one or two-letter abbreviation
    for the given directional. This follows USPS addressing standards.
}
\description{
Convert directionals to the USPS preferred abbreviations
}
