% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\name{pm_streetSuf_all}
\alias{pm_streetSuf_all}
\title{Does Street Suffix Dictionary Return a Match for All Observations}
\usage{
pm_streetSuf_all(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for street suffixes. If none is specified, the full default
street suffix dictionary will be used.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains a street suffix
    name or abbreviation for every observation in the data set and \code{FALSE} otherwise.
}
\description{
Determine whether the street suffix dictionary returns matches for all observations.
}
