% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\name{pm_streetSuf_any}
\alias{pm_streetSuf_any}
\title{Does Street Suffix Dictionary Return Any Matches}
\usage{
pm_streetSuf_any(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for street suffixes}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A logical scalar is returned that is \code{TRUE} if the data contains at
    least one street suffix name or abbrevation in the given dictionary and \code{FALSE}
    if they do not.
}
\description{
Determine whether the street suffix dictionary returns any matches.
}
