% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\name{pm_streetSuf_none}
\alias{pm_streetSuf_none}
\title{Return Only Unmatched Observations From pm_streetSuf_detect}
\usage{
pm_streetSuf_none(.data, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{dictionary}{Optional; a tbl created with \code{pm_dictionary} to be used
as a master list for street suffixes. If none is specified, the full default
street suffix dictionary will be used.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\description{
Automatically subset the results of \link{pm_streetSuf_detect} to
   return only observations that were not found in the dictionary.
}
