% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\name{pm_streetSuf_std}
\alias{pm_streetSuf_std}
\title{Standardize Parsed Street Suffixes}
\usage{
pm_streetSuf_std(.data, var, dictionary, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{var}{A character variable that may contain street suffixes}

\item{dictionary}{A tbl created with \code{pm_dictionary} to be used
as a master list for street suffixes.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\description{
Convert street suffixes to USPS preferred abbreviation.
}
