% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_stSuffix.R
\name{pm_stSuffix}
\alias{pm_stSuffix}
\title{Parse Street Suffixes}
\usage{
pm_stSuffix(.data, stName, std = TRUE, output)
}
\arguments{
\item{.data}{A tbl}

\item{std}{A logical scalar. Should suffix types be standardized?}

\item{output}{Optional name for output variable}

\item{address}{Name of the street address variable containing suffix types}
}
\description{
Parse Street Suffixes
}
