% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_stSuffixDir.R
\name{pm_stSuffixDir}
\alias{pm_stSuffixDir}
\title{Parse Suffix Direction}
\usage{
pm_stSuffixDir(.data, stName, std = TRUE, output)
}
\arguments{
\item{.data}{A tbl}

\item{std}{A logical scalar. Should directional values be standardized?}

\item{output}{Optional name for output variable}

\item{address}{Name of the street address variable containing suffix directions}
}
\description{
\code{pm_stSuffixDir} identifies streets with suffix directions (i.e. 123 Washington St NW) and
parses those directions into their own variable.
}
