% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{pm_type_partial}
\alias{pm_type_partial}
\title{Return Only Partial Records}
\usage{
pm_type_partial(.data)
}
\arguments{
\item{.data}{A source tibble that has already had identification
numbers added using \link{pm_identify}.}
}
\value{
A tibble with only observations that are \code{"unknown"} in \code{pm.type}.
}
\description{
Subsets addresses that are marked as \code{"partial"} in \code{pm.type} so that
   they can be individually evaluated and updated prior to standardization.
}
