% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_word2num.R
\name{pm_word2num}
\alias{pm_word2num}
\title{Convert Words to Numbers}
\usage{
pm_word2num(word)
}
\arguments{
\item{word}{A character string to be converted}
}
\value{
A numeric value
}
\description{
\code{pm_word2num} converts words (e.g. "one" and "first") to numeric values (e.g. "1").
}
