% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdSuffixTbl.R
\docType{data}
\name{stdSuffixTbl}
\alias{stdSuffixTbl}
\title{Standardized Street Suffixes}
\format{A data frame with 502 rows and 3 variables:
\describe{
  \item{suf_pri}{full category name}
  \item{suf_com}{commonly accepted alternatives}
  \item{suf_std}{standard abbreviation}
  }}
\usage{
stdSuffixTbl
}
\description{
A data set containing standardized street suffixes and accepted alternatives from the USPS.
}
\keyword{datasets}
