% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushi.R
\docType{data}
\name{sushi1}
\alias{sushi1}
\title{Sushi Resturants in the St. Louis Region}
\format{A tibble with 30 rows and 3 variables:
\describe{
  \item{name}{resturant name}
  \item{address}{street address include city, state, and postal code}
  \item{visit}{visit date}
}}
\usage{
data(sushi1)
}
\description{
A selection of sushi resturants in the St. Louis, Missouri region.
   These data are in long form, with multiple visits in a single year to some
   of the resturants. They also contain a degree of messiness, with a mix of
   upper and lower case tying, some street types abbreviated (i.e. "St") and
   others entered in full (i.e. "Street"), and an example of two ways in which
   unit numbers may appear. State names are given both in full and using the
   standard two-letter abbreviation, and city names as well as postal codes
   are included. Postal codes are given in both the standard five digit
   form as well as the elongated "zip+4" format.
}
\examples{
str(sushi1)
head(sushi1)

}
\keyword{datasets}
