% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushi.R
\docType{data}
\name{sushi2}
\alias{sushi2}
\title{Sushi Resturants in the Ciy of St. Louis}
\format{A tibble with 19 rows and 3 variables:
\describe{
  \item{name}{resturant name}
  \item{address}{street address}
  \item{visit}{visit date}
}}
\usage{
data(sushi2)
}
\description{
A selection of sushi resturants in the St. Louis, Missouri region.
   These data are in long form, with multiple visits in a single year to some
   of the resturants. They also contain a degree of messiness, with a mix of
   upper and lower case tying, some street types abbreviated (i.e. "St") and
   others entered in full (i.e. "Street"), and an example of two ways in which
   unit numbers may appear. Unlike \code{sushi1}, these data do not contain
   city, state, or postal code data.
}
\examples{
str(sushi2)
head(sushi2)

}
\keyword{datasets}
