% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{pm_append}
\alias{pm_append}
\title{Create Dictionary Object Appendicies}
\usage{
pm_append(type, input, output, locale = "us")
}
\arguments{
\item{type}{A string indicating the grammatical address element the dictionary
should represent. Current options are \code{"state"}, \code{"city"},
\code{"street"}, \code{"directional"}, and \code{"suffix"}.}

\item{input}{A character scalar or vector containing possible terms existing in
the data. This should be the same length as \code{output}.}

\item{output}{A character scalar or vector containing desired output for each input.
This should be the same length as \code{input}. You may use \code{NA} values
in the \code{output} vector for inputs that already are in the preferred,
standardized form. This argument is optional when \code{type} is equal to
\code{"city"}.}

\item{locale}{A string indicating the country these data represent; the only
current option is \code{"us"} but this is included to facilitate future expansion.}
}
\value{
A \code{postmastr} dictionary object, which will always include an input column
    of possible terms for the given grammatical address element. All dictionary objects
    except for city dictionaries will also contain an output column with the
    preferred output for each input. For city dictionaries, the \code{output} argument
    is optional.

    For American addresses, these outputs should follow
    United States Postal Service guidelines. These dictionary objects can be used as
    stand-alone dictionaries or fed into a \code{\link{pm_dictionary}} call as the input
    for the \code{append} parameter.
}
\description{
This function allows for the creation of dictionary objects that are
    either optional or required elements of other \code{postmastr} functions. These
    objects are appendicies that can serve either as stand-alone dictionaries or
    used in a \code{\link{pm_dictionary}} call as the input  for the \code{append}
    parameter.

    Once created, dictionary objects should not be modified as the formatting of dictionary
    elements provides other \code{postmastr} functions with predictable inputs. The
    \code{"state"}, \code{"directional"}, and \code{"suffix"} dictionaries are based on
    tables that are also exported so that they can be previewed and tested.
}
\examples{
# create stand-alone state dictionary
mo <- pm_append(type = "state", input = c("Missouri", "MO", "MISSOURI"),
    output = c("MO", "MO", "MO"), locale = "us")

# add custom abbreviation for Massachusetts to state dictionary
## create dictionary appendix
ma <- pm_append(type = "state", input = "Mass", output = "MA", locale = "us")

## add dictionary appendix to dictionary
pm_dictionary(type = "state", filter = c("CT", "MA", "ME", "NH", "RI", "VT"),
    append = ma, case = "title", locale = "us")

# create stand-alone city dictionary, with spelling correction for some cities
cities <- pm_append(type = "city",
    input = c("Brentwood", "Clayton", "CLAYTON", "Maplewood", "St. Louis",
              "SAINT LOUIS", "Webster Groves"),
    output = c(NA, NA, "Clayton", NA, NA, "St. Louis", NA),
    locale = "us")

# create stand-alone street dictionary
cities <- pm_append(type = "street",
    input = c("MLK", "M.L.K."),
    output = c("Dr. Martin Luther King, Jr.", "Dr. Martin Luther King, Jr."),
    locale = "us")

}
\seealso{
\code{pm_dictionary}
}
