% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{pm_identify}
\alias{pm_identify}
\title{Identify Observations}
\usage{
pm_identify(.data, var)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable containing address data to be parsed}
}
\value{
A tibble with both the \code{pm.id} and \code{pm.uid} variables added
   in the first two positions of the data set.
}
\description{
Adds two identification numbers to raw data that will be used
    for matching parsed data with the original, raw data once parsing
    is complete. One (\code{pm.id}) uniquely identifies each observation,
    while the other uniquely identifies each distinct street address
    (\code{pm.uid}).
}
\details{
\code{postmastr} functions are designed to operate
    on unique street addresses rather than on an entire data set to increase
    speed and performance. The \code{pm.uid} number helps facilitate the
    matching process between processed and original data while the observation
    identification number \code{pm.id} preserves the original sort order of
    the data. These variable names should not be changed - subsequent functions
    applied to the prepared data check for their presence before executing
    to ensure that they remain, and will error if they are not found.
}
