% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{pm_isState}
\alias{pm_isState}
\title{Detect Presence of State Name or Abbreviation}
\usage{
pm_isState(.data, var, directory)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable that may contain city names}
}
\value{
A tibble with a new logical variable \code{pm.isState} that is
    \code{TRUE} if a state name or abbreviation is found in the address
    and \code{FALSE} otherwise.
}
\description{
Determine the presence of state names or abbreviations
    in a string.
}
