% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{pm_mutate}
\alias{pm_mutate}
\title{Modify Improperly Parsed Addresses}
\usage{
pm_mutate(.data, pm.uid, pm.house, pm.houseRange, pm.houseFrac,
  pm.houseSuf, pm.preDir, pm.street, pm.streetSuf, pm.sufDir, pm.unitType,
  pm.unitNum, pm.city, pm.state, pm.zip, pm.zip4)
}
\description{
Difficult to parse addresses, particularly those that have
   alphanumeric house ranges as well as units associated with them,
   may be mis-parsed by \code{postmastr}. Once all parsing has been
   completed, but before the data are rebuilt, this function can be used
   to manually fix mis-parsed address data.
}
