% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pm_parse}
\alias{pm_parse}
\title{Parse Street Address}
\usage{
pm_parse(.data, style, locale = "us", newVar, keep_parsed = FALSE, keep_ids = FALSE,
    dirDict, streetDict, suffixDict, unitDict, cityDict, stateDict)
}
\arguments{
\item{.data}{A source data set to be parsed}

\item{style}{One of either \code{"full"} or \code{"short"}. A short address contains,
at the most, a house number, street directionals, a street name, a street suffix,
and a unit type and number. A full address contains all of the selements of a short
address as well as a, at the most, a city, state, and postal code.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}

\item{newVar}{Name of new variable to store rebuilt address in.}

\item{keep_parsed}{Logical scalar; if \code{TRUE}, all parsed elements will be
added to the source data after replacement. Otherwise, if \code{FALSE},
only the rebuilt address will be added to the source data (default).}

\item{keep_ids}{Logical scalar; if \code{TRUE}, the identification numbers
will be kept in the source data after replacement. Otherwise, if \code{FALSE},
they will be removed (default).}

\item{dirDict}{Optional; name of directional dictionary object}

\item{streetDict}{Optional; name of street dictionary object}

\item{suffixDict}{Optional; name of street suffix dictionary object}

\item{unitDict}{Optional; name of unit dictionary object}

\item{cityDict}{Optional; name of city dictionary object}

\item{stateDict}{Optional; name of state dictionary object}
}
\description{
A wrapper around the parse functions that can be used to shorten all
    of \code{postmastr}'s code down to a single function call once dictionaries
    have been created and tested against the data.
}
