% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pm_parse}
\alias{pm_parse}
\title{Parse Street Address}
\usage{
pm_parse(.data, input, var, output, newVar, ordinal = TRUE, include_commas = FALSE,
    include_unit = TRUE, keep_parsed = "no", keep_ids = FALSE, dirDict, streetDict,
    suffixDict, unitDict, cityDict, stateDict, locale = "us")
}
\arguments{
\item{.data}{A source data set to be parsed}

\item{input}{Describes the format of the source address. One of either \code{"full"} or \code{"short"}.
A short address contains, at the most, a house number, street directionals, a street name,
a street suffix, and a unit type and number. A full address contains all of the selements of a
short address as well as a, at the most, a city, state, and postal code.}

\item{var}{A character variable containing address data to be parsed}

\item{output}{Describes the format of the output address. One of either \code{"full"} or \code{"short"}.
A short address contains, at the most, a house number, street directionals, a street name,
a street suffix, and a unit type and number. A full address contains all of the selements of a
short address as well as a, at the most, a city, state, and postal code.}

\item{newVar}{Name of new variable to store rebuilt address in.}

\item{ordinal}{A logical scalar; if \code{TRUE}, street names that contain numeric words values
(i.e. "Second") will be converted and standardized to ordinal values (i.e. "2nd"). The
default is \code{TRUE} because it returns much more compact clean addresses (i.e.
"168th St" as opposed to "One Hundred Sixty Eigth St").}

\item{include_commas}{A logical scalar; if \code{TRUE}, a comma is added both before and after the city
name in rebuild addresses. If \code{FALSE} (default), no punctuation is added.}

\item{include_unit}{A logical scalar; if \code{TRUE} (default), the unit name and number (if given)
will be included in the output string. Otherwise if \code{FALSE}, the unit name and number
will not be included.}

\item{keep_parsed}{Character string; if \code{"yes"}, all parsed elements will be
added to the source data after replacement. If \code{"limited"}, only the \code{pm.city},
\code{pm.state}, and postal code variables will be retained. Otherwise, if \code{"no"},
only the rebuilt address will be added to the source data (default).}

\item{keep_ids}{Logical scalar; if \code{TRUE}, the identification numbers
will be kept in the source data after replacement. Otherwise, if \code{FALSE},
they will be removed (default).}

\item{dirDict}{Optional; name of directional dictionary object}

\item{streetDict}{Optional; name of street dictionary object}

\item{suffixDict}{Optional; name of street suffix dictionary object}

\item{unitDict}{Optional; name of unit dictionary object}

\item{cityDict}{Optional; name of city dictionary object}

\item{stateDict}{Optional; name of state dictionary object}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\description{
A wrapper around the parse functions that can be used to shorten all
    of \code{postmastr}'s code down to a single function call once dictionaries
    have been created and tested against the data.
}
