% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{pm_prep}
\alias{pm_prep}
\title{Prepare Data}
\usage{
pm_prep(.data, var)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{A character variable containing address data to be parsed}
}
\description{
Prepare Data
}
