% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{pm_prep}
\alias{pm_prep}
\title{Create postmastr Object for Parsing}
\usage{
pm_prep(.data, var)
}
\arguments{
\item{.data}{A tibble that has already had identification numbers added
using \link{pm_identify}.}

\item{var}{A character variable containing address data to be parsed}
}
\description{
Once identification numbers have been been added, this
   function creates a subset that will be parsed and then re-applied
   to the primary data set.
}
