% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pm_rebuild}
\alias{pm_rebuild}
\title{Rebuild Street Address}
\usage{
pm_rebuild(.data, start, end, include_commas = FALSE, locale = "us")
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep}}

\item{start}{Variable name to begin rebuilding process with, typically the house number}

\item{end}{Variable name to end rebuilding process with, typically the street suffix or postal code}

\item{include_commas}{A logical scalar; if \code{TRUE}, a comma is added both before and after the city
name in rebuild addresses. If \code{FALSE} (default), no punctuation is added.}

\item{locale}{A string indicating the country these data represent; the only
current option is "us" but this is included to facilitate future expansion.}
}
\value{
A copy of the pomstmastr object with a standardized address variable.
}
\description{
Create a single address from parsed components. Data are automatically re-ordered
    as part of this process, so there is no need to subset columns prior to executing
    this function.
}
