% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pm_replace}
\alias{pm_replace}
\title{Add Address to Source Data}
\usage{
pm_replace(.data, source, newVar, keep_parsed, keep_ids = FALSE)
}
\arguments{
\item{.data}{A postmastr object created with \link{pm_prep} that has been readied
for replacement by (a) fully parsing the data and (b) rebuilding a sinle
address field.}

\item{source}{Original source data to merge clean addresses with.}

\item{newVar}{Name of new variable to store rebuilt address in.}

\item{keep_parsed}{Character string; if \code{"yes"}, all parsed elements will be
added to the source data after replacement. If \code{"limited"}, only the \code{pm.city},
\code{pm.state}, and postal code variables will be retained. Otherwise, if \code{"no"},
only the rebuilt address will be added to the source data (default).}

\item{keep_ids}{Logical scalar; if \code{TRUE}, the identification numbers
will be kept in the source data after replacement. Otherwise, if \code{FALSE},
they will be removed (default).}
}
\value{
The source data with a standardized address variable and, optionally, parsed
    address elements.
}
\description{
Adds standardized address and, optionally, parsed elements back into
    original source data frame.
}
