% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pm_replace}
\alias{pm_replace}
\title{Add Parsed Street Address Data to Source Data Set}
\usage{
pm_replace(source, street, intersect, operator = "at", unnest = FALSE)
}
\arguments{
\item{source}{Original source data to merge clean addresses with.}

\item{street}{A postmastr object created with \link{pm_prep} with street addresses that has been readied
for replacement by fully parsing the data.}

\item{intersect}{A postmastr object created with \link{pm_prep} with intersections that has been readied
for replacement by fully parsing the data.}

\item{operator}{A character scalar to be used as the intersection operator (between the 'x' and 'y' sides
of the intersection).}

\item{unnest}{A logical scalar; if \code{TRUE}, house ranges will be unnested (i.e. a house range that
has been expanded to cover four addresses with \code{\link{pm_houseRange_parse}} will be converted
from a single observation to four observations, one for each house number). If \code{FALSE} (default),
the single observation will remain.}
}
\value{
The source data with a parsed address elements added to the left side of the source data.
}
\description{
Adds standardized address and, optionally, unnest house ranges. All logical variables
   that were created during the data cleaning process (i.e. \code{pm.hasHouse}, \code{pm.hasDir}, etc.)
   are removed at this stage.
}
